添加根据用户id获取店铺信息

+++ litemall-db/src/main/java/org/linlinjava/litemall/db/dao/MallMapper.java
@@ -1,18 +1,17 @@
 package org.linlinjava.litemall.db.dao;
 
 
-import org.apache.ibatis.annotations.Mapper;
 import org.linlinjava.litemall.db.chao.domain.Mall;
 
 import java.util.List;
 
 
 public interface MallMapper {
-    int deleteByPrimaryKey(Byte id);
+    int deleteByPrimaryKey(int id);
 
     int insert(Mall record);
 
-    Mall selectByPrimaryKey(Byte id);
+    Mall selectByPrimaryKey(int id);
 
     List<Mall> selectAll();
